src/requirements.txt
google-cloud-storage
google-cloud-language
vertexai
nltk
kfp
rouge-score